Update the renderers
