+++ packages/core/src/renderer/Line.ts
@@ -1,8 +1,106 @@
-import { IFloatV, IColorV, IStrV } from "types/value";
-import { toSvgPaintProperty, toSvgOpacityProperty, toScreen } from "utils/Util";
-import { arrowHead, makeRoomForArrows } from "./Arrow";
+import { Shape } from "types/shape";
+import { IFloatV, IColorV, IStrV, IVectorV } from "types/value";
+import {
+  toSvgPaintProperty,
+  toSvgOpacityProperty,
+  toScreen,
+  arrowheads,
+  round2,
+} from "utils/Util";
 import { attrAutoFillSvg, attrTitle, DASH_ARRAY } from "./AttrHelper";
 import { ShapeProps } from "./Renderer";
+
+export const arrowHead = (
+  id: string,
+  color: string,
+  opacity: number,
+  style: string,
+  size: number
+): SVGMarkerElement => {
+  const arrow = arrowheads[style];
+
+  const marker = document.createElementNS(
+    "http://www.w3.org/2000/svg",
+    "marker"
+  );
+  marker.setAttribute("id", id);
+  marker.setAttribute("markerUnits", "strokeWidth");
+  marker.setAttribute("markerWidth", round2(arrow.width * size).toString());
+  marker.setAttribute("markerHeight", round2(arrow.height * size).toString());
+  marker.setAttribute("viewBox", arrow.viewbox);
+  marker.setAttribute("refX", arrow.refX.toString());
+  marker.setAttribute("refY", arrow.refY.toString());
+  marker.setAttribute("orient", "auto-start-reverse");
+
+  const path = document.createElementNS("http://www.w3.org/2000/svg", "path");
+  path.setAttribute("d", arrow.path);
+  path.setAttribute("fill", color);
+  path.setAttribute("fill-opacity", opacity.toString());
+  marker.appendChild(path);
+
+  return marker;
+};
+
+const makeRoomForArrows = (shape: Shape): [number[][], string[]] => {
+  // Keep a list of which input properties we programatically mapped
+  const attrMapped: string[] = [];
+
+  const [lineSX, lineSY] = (shape.properties.start as IVectorV<number>)
+    .contents as [number, number];
+  const [lineEX, lineEY] = (shape.properties.end as IVectorV<number>)
+    .contents as [number, number];
+
+  const arrowheadStyle = (shape.properties.arrowheadStyle as IStrV).contents;
+  const arrowheadSize = (shape.properties.arrowheadSize as IFloatV<number>)
+    .contents;
+  const thickness = (shape.properties.strokeWidth as IFloatV<number>).contents;
+  attrMapped.push(
+    "start",
+    "end",
+    "arrowheadStyle",
+    "arrowheadSize",
+    "strokeWidth"
+  );
+
+  // height * size = Penrose computed arrow size
+  // multiplied by thickness since the arrow size uses markerUnits, which is strokeWidth by default:
+  // https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/markerUnits
+  const arrowHeight =
+    arrowheads[arrowheadStyle].height * arrowheadSize * thickness;
+  const length = Math.sqrt((lineSX - lineEX) ** 2 + (lineSY - lineEY) ** 2);
+
+  // Subtract off the arrowHeight from each side.
+  // See https://math.stackexchange.com/a/2045181 for a derivation.
+  let arrowSX, arrowSY;
+  if (shape.properties.leftArrowhead.contents) {
+    [arrowSX, arrowSY] = [
+      lineSX - (arrowHeight / length) * (lineSX - lineEX),
+      lineSY - (arrowHeight / length) * (lineSY - lineEY),
+    ];
+  } else {
+    [arrowSX, arrowSY] = [lineSX, lineSY];
+  }
+  attrMapped.push("startArrowhead");
+
+  let arrowEX, arrowEY;
+  if (shape.properties.rightArrowhead.contents) {
+    [arrowEX, arrowEY] = [
+      lineEX - (arrowHeight / length) * (lineEX - lineSX),
+      lineEY - (arrowHeight / length) * (lineEY - lineSY),
+    ];
+  } else {
+    [arrowEX, arrowEY] = [lineEX, lineEY];
+  }
+  attrMapped.push("endArrowhead");
+
+  return [
+    [
+      [arrowSX, arrowSY],
+      [arrowEX, arrowEY],
+    ],
+    attrMapped,
+  ];
+};
 
 const Line = ({ shape, canvasSize }: ShapeProps): SVGGElement => {
   const [
@@ -13,11 +111,11 @@ const Line = ({ shape, canvasSize }: ShapeProps): SVGGElement => {
   const [ex, ey] = toScreen([arrowEX, arrowEY], canvasSize);
   const path = `M ${sx} ${sy} L ${ex} ${ey}`;
   const color = toSvgPaintProperty(
-    (shape.properties.color as IColorV<number>).contents
+    (shape.properties.strokeColor as IColorV<number>).contents
   );
-  const thickness = (shape.properties.thickness as IFloatV<number>).contents;
+  const thickness = (shape.properties.strokeWidth as IFloatV<number>).contents;
   const opacity = toSvgOpacityProperty(
-    (shape.properties.color as IColorV<number>).contents
+    (shape.properties.strokeColor as IColorV<number>).contents
   );
   const leftArrowId = shape.properties.name.contents + "-leftArrowhead";
   const rightArrowId = shape.properties.name.contents + "-rightArrowhead";
@@ -35,8 +133,8 @@ const Line = ({ shape, canvasSize }: ShapeProps): SVGGElement => {
     arrowHead(rightArrowId, color, opacity, arrowheadStyle, arrowheadSize)
   );
   attrToNotAutoMap.push(
-    "color",
-    "thickness",
+    "strokeColor",
+    "strokeWidth",
     "arrowheadStyle",
     "arrowheadSize"
   );
@@ -80,14 +178,14 @@ const Line = ({ shape, canvasSize }: ShapeProps): SVGGElement => {
   }
   attrToNotAutoMap.push("strokeLineCap");
 
-  // TODO: dedup in AttrHelper (problem: thickness vs strokeWidth)
-  if (shape.properties.leftArrowhead.contents === true) {
+  // TODO: dedup in AttrHelper
+  if (shape.properties.startArrowhead.contents === true) {
     pathElem.setAttribute("marker-start", `url(#${leftArrowId})`);
-    attrToNotAutoMap.push("leftArrowhead");
+    attrToNotAutoMap.push("startArrowhead");
   }
-  if (shape.properties.rightArrowhead.contents === true) {
+  if (shape.properties.endArrowhead.contents === true) {
     pathElem.setAttribute("marker-end", `url(#${rightArrowId})`);
-    attrToNotAutoMap.push("rightArrowhead");
+    attrToNotAutoMap.push("endArrowhead");
   }
   elem.appendChild(pathElem);
   attrToNotAutoMap.push(...attrTitle(shape, elem));
