+++ packages/core/src/renderer/Polygon.ts
@@ -2,7 +2,6 @@ import {
   attrAutoFillSvg,
   attrFill,
   attrPoints,
-  attrPolyCenter,
   attrScale,
   attrStroke,
   attrTitle,
@@ -23,7 +22,6 @@ const Polygon = ({ shape, canvasSize }: ShapeProps): SVGPolygonElement => {
   attrToNotAutoMap.push(...attrStroke(shape, elem));
   attrToNotAutoMap.push(...attrTitle(shape, elem));
   attrToNotAutoMap.push(...attrPoints(shape, elem));
-  attrToNotAutoMap.push(...attrPolyCenter(shape, canvasSize, elem));
   attrToNotAutoMap.push(...attrScale(shape, elem));
 
   // Directrly Map across any "unknown" SVG properties
