+++ packages/core/src/renderer/Polyline.ts
@@ -1,7 +1,6 @@
 import {
   attrPoints,
   attrScale,
-  attrPolyCenter,
   attrStroke,
   attrTitle,
   attrAutoFillSvg,
@@ -23,7 +22,6 @@ const Polyline = ({ shape, canvasSize }: ShapeProps): SVGPolylineElement => {
   attrToNotAutoMap.push(...attrStroke(shape, elem));
   attrToNotAutoMap.push(...attrTitle(shape, elem));
   attrToNotAutoMap.push(...attrPoints(shape, elem));
-  attrToNotAutoMap.push(...attrPolyCenter(shape, canvasSize, elem));
   attrToNotAutoMap.push(...attrScale(shape, elem));
 
   // Directrly Map across any "unknown" SVG properties
